+++ modules/core/src/controllers/orthographic-controller.js
@@ -1,14 +1,5 @@
 import Controller from './controller';
 import {OrbitState} from './orbit-controller';
-import LinearInterpolator from '../transitions/linear-interpolator';
-import {TRANSITION_EVENTS} from './transition-manager';
-
-const LINEAR_TRANSITION_PROPS = {
-  transitionDuration: 300,
-  transitionEasing: t => t,
-  transitionInterpolator: new LinearInterpolator(['target', 'zoom']),
-  transitionInterruption: TRANSITION_EVENTS.BREAK
-};
 
 export default class OrthographicController extends Controller {
   constructor(props) {
@@ -21,8 +12,7 @@ export default class OrthographicController extends Controller {
     return false;
   }
 
-  _getTransitionProps() {
-    // Enables Transitions on double-tap and key-down events.
-    return LINEAR_TRANSITION_PROPS;
+  get linearTransitionProps() {
+    return ['target', 'zoom'];
   }
 }