+++ modules/core/src/views/orthographic-view.js
@@ -2,6 +2,8 @@ import View from './view';
 import Viewport from '../viewports/viewport';
 
 import {Matrix4} from 'math.gl';
+import {pixelsToWorld} from '@math.gl/web-mercator';
+import * as vec2 from 'gl-matrix/vec2';
 import OrthographicController from '../controllers/orthographic-controller';
 
 const viewMatrix = new Matrix4().lookAt({eye: [0, 0, 1]});
@@ -43,6 +45,16 @@ class OrthographicViewport extends Viewport {
       projectionMatrix: getProjectionMatrix({width, height, near, far}),
       zoom
     });
+  }
+
+  /* Needed by LinearInterpolator */
+  panByPosition(coords, pixel) {
+    const fromLocation = pixelsToWorld(pixel, this.pixelUnprojectionMatrix);
+
+    const translate = vec2.add([], coords, vec2.negate([], fromLocation));
+    const newCenter = vec2.add([], this.center, translate);
+
+    return {target: newCenter};
   }
 }