+++ modules/geo-layers/src/tile-3d-layer/tile-3d-layer.js
@@ -2,7 +2,8 @@ import GL from '@luma.gl/constants';
 import {Geometry} from '@luma.gl/core';
 import {COORDINATE_SYSTEM, CompositeLayer} from '@deck.gl/core';
 import {PointCloudLayer} from '@deck.gl/layers';
-import {ScenegraphLayer, SimpleMeshLayer} from '@deck.gl/mesh-layers';
+import {ScenegraphLayer} from '@deck.gl/mesh-layers';
+import {default as _MeshLayer} from '../mesh-layer/mesh-layer';
 import {log} from '@deck.gl/core';
 
 import {load} from '@loaders.gl/core';
@@ -22,7 +23,7 @@ const defaultProps = {
   onTileLoad: {type: 'function', value: tileHeader => {}, compare: false},
   onTileUnload: {type: 'function', value: tileHeader => {}, compare: false},
   onTileError: {type: 'function', value: (tile, message, url) => {}, compare: false},
-  getSimpleMeshLayerColor: {type: 'function', value: tileHeader => [255, 255, 255], compare: false}
+  _getMeshColor: {type: 'function', value: tileHeader => [255, 255, 255], compare: false}
 };
 
 export default class Tile3DLayer extends CompositeLayer {
@@ -33,7 +34,9 @@ export default class Tile3DLayer extends CompositeLayer {
     // prop verification
     this.state = {
       layerMap: {},
-      tileset3d: null
+      tileset3d: null,
+      activeViewports: {},
+      lastUpdatedViewports: null
     };
   }
 
@@ -52,14 +55,31 @@ export default class Tile3DLayer extends CompositeLayer {
     }
 
     if (changeFlags.viewportChanged) {
-      const {tileset3d} = this.state;
-      this._updateTileset(tileset3d);
+      const {activeViewports} = this.state;
+      const viewportsNumber = Object.keys(activeViewports).length;
+      if (viewportsNumber) {
+        this._updateTileset(activeViewports);
+        this.state.lastUpdatedViewports = activeViewports;
+        this.state.activeViewports = {};
+      }
     }
     if (changeFlags.propsChanged) {
       const {layerMap} = this.state;
       for (const key in layerMap) {
         layerMap[key].needsUpdate = true;
       }
+    }
+  }
+
+  activateViewport(viewport) {
+    const {activeViewports, lastUpdatedViewports} = this.state;
+    this.internalState.viewport = viewport;
+
+    activeViewports[viewport.id] = viewport;
+    const lastViewport = lastUpdatedViewports?.[viewport.id];
+    if (!lastViewport || !viewport.equals(lastViewport)) {
+      this.setChangeFlags({viewportChanged: true});
+      this.setNeedsUpdate();
     }
   }
 
@@ -110,13 +130,14 @@ export default class Tile3DLayer extends CompositeLayer {
       layerMap: {}
     });
 
-    this._updateTileset(tileset3d);
+    this._updateTileset(this.state.activeViewports);
     this.props.onTilesetLoad(tileset3d);
   }
 
   _onTileLoad(tileHeader) {
+    const {lastUpdatedViewports} = this.state;
     this.props.onTileLoad(tileHeader);
-    this._updateTileset(this.state.tileset3d);
+    this._updateTileset(lastUpdatedViewports);
     this.setNeedsUpdate();
   }
 
@@ -126,12 +147,14 @@ export default class Tile3DLayer extends CompositeLayer {
     this.props.onTileUnload(tileHeader);
   }
 
-  _updateTileset(tileset3d) {
-    const {timeline, viewport} = this.context;
-    if (!timeline || !viewport || !tileset3d) {
+  _updateTileset(viewports) {
+    const {tileset3d} = this.state;
+    const {timeline} = this.context;
+    const viewportsNumber = Object.keys(viewports).length;
+    if (!timeline || !viewportsNumber || !tileset3d) {
       return;
     }
-    const frameNumber = tileset3d.update(viewport);
+    const frameNumber = tileset3d.update(Object.values(viewports));
     const tilesetChanged = this.state.frameNumber !== frameNumber;
     if (tilesetChanged) {
       this.setState({frameNumber});
@@ -228,17 +251,18 @@ export default class Tile3DLayer extends CompositeLayer {
 
   _makeSimpleMeshLayer(tileHeader, oldLayer) {
     const content = tileHeader.content;
-    const {attributes, modelMatrix, cartographicOrigin, texture} = content;
-    const {getSimpleMeshLayerColor} = this.props;
+    const {attributes, indices, modelMatrix, cartographicOrigin, material, featureIds} = content;
+    const {_getMeshColor} = this.props;
 
     const geometry =
       (oldLayer && oldLayer.props.mesh) ||
       new Geometry({
         drawMode: GL.TRIANGLES,
-        attributes: getMeshGeometry(attributes)
+        attributes: getMeshGeometry(attributes),
+        indices
       });
 
-    const SubLayerClass = this.getSubLayerClass('mesh', SimpleMeshLayer);
+    const SubLayerClass = this.getSubLayerClass('mesh', _MeshLayer);
 
     return new SubLayerClass(
       this.getSubLayerProps({
@@ -248,12 +272,12 @@ export default class Tile3DLayer extends CompositeLayer {
         id: `${this.id}-mesh-${tileHeader.id}`,
         mesh: geometry,
         data: SINGLE_DATA,
-        getPosition: [0, 0, 0],
-        getColor: getSimpleMeshLayerColor(tileHeader),
-        texture,
+        getColor: _getMeshColor(tileHeader),
+        pbrMaterial: material,
         modelMatrix,
         coordinateOrigin: cartographicOrigin,
-        coordinateSystem: COORDINATE_SYSTEM.METER_OFFSETS
+        coordinateSystem: COORDINATE_SYSTEM.METER_OFFSETS,
+        featureIds
       }
     );
   }
@@ -305,6 +329,12 @@ function getMeshGeometry(contentAttributes) {
   }
   if (contentAttributes.texCoords) {
     attributes.texCoords = contentAttributes.texCoords;
+  }
+  if (contentAttributes.colors) {
+    attributes.colors = contentAttributes.colors;
+  }
+  if (contentAttributes.uvRegions) {
+    attributes.uvRegions = contentAttributes.uvRegions;
   }
   return attributes;
 }