+++ test/modules/geo-layers/tile-3d-layer/tile-3d-layer.spec.js
@@ -23,28 +23,12 @@ import {testLayerAsync} from '@deck.gl/test-utils';
 import {Tile3DLayer} from '@deck.gl/geo-layers';
 import {WebMercatorViewport} from '@deck.gl/core';
 
-const fetchFile = url => {
-  url = url
-    .replace(/\?.+$/, '') // strip query parameters
-    .replace(/^\//, './test/data/3d-tiles/');
-  return require('fs').readFileSync(url);
-};
-
 test('Tile3DLayer', async t => {
-  let oldFetch;
-  /* global global, Response */
-  const needsFetchPolyfill = typeof Response === 'undefined';
-  if (needsFetchPolyfill) {
-    oldFetch = global.fetch;
-    global.fetch = fetchFile;
-  }
-
   const testCases = [
     {
       props: {
         data: './test/data/3d-tiles/tileset.json',
-        getPointColor: [0, 0, 0],
-        loadOptions: needsFetchPolyfill ? {'3d-tiles': {isTileset: true}} : {}
+        getPointColor: [0, 0, 0]
       },
       onBeforeUpdate: () => t.comment('inital load'),
       onAfterUpdate: ({layer, subLayers}) => {
@@ -79,8 +63,5 @@ test('Tile3DLayer', async t => {
     onError: t.notOk
   });
 
-  if (oldFetch) {
-    global.fetch = oldFetch;
-  }
   t.end();
 });