+++ CHANGELOG.md
@@ -21,6 +21,25 @@ Ref: http://keepachangelog.com/en/0.3.0/
 
 ### deck.gl v8.5 Prereleases
 
+#### deck.gl [8.5.0-alpha.8] - Jun 5 2021
+
+- Pass through loadOptions and AbortSignal to MVTLayer & TerrainLayer (#5837)
+- Refactor OrbitController (#5825)
+- Allow CompositeLayer to filter sub layers during redraw (#5820)
+- Add getTilerColor method to tile-3d-layer (#5814)
+- Refactor linear transition in controllers (#5824)
+- i3s: support vertex colors (#5756)
+- i3s: support uvRegions (#5760)
+- Fix heatmap colorDomain (#5802)
+- feat(geo-layers): Private MeshLayer with pbr material support (I3S) (#5761)
+
+#### deck.gl [8.5.0-alpha.7] - May 25 2021
+
+- geo: i3s - support indices (#5807)
+- Bump @loaders.gl to 3.0.0-alpha.16 to pick up earcut migration to math.gl (#5805)
+- Set default of MVTLayer binary to true (#5800)
+- Fix size projection in billboard mode (#5798)
+
 #### deck.gl [8.5.0-alpha.6] - May 17 2021
 
 - Bump @loaders.gl to 3.0.0-alpha.14 (#5780)